# src/docs/swagger.yaml
openapi: 3.0.3
info:
  title: NovaHouse Chatbot API
  description: |
    API dla inteligentnego chatbota NovaHouse - firmy specjalizującej się w kompleksowych remontach i wykończeniach wnętrz.

    ## Features
    - 🤖 AI-powered chatbot z Google Gemini
    - 📊 Advanced analytics & reporting
    - 📧 Email notifications
    - 📅 Booksy integration
    - 🔗 Monday.com CRM sync
    - 🔒 RODO/GDPR compliance
    - 🛡️ Rate limiting & security
    - 💾 Caching system

  version: 2.1.0
  contact:
    name: NovaHouse Support
    url: https://novahouse.pl
    email: kontakt@novahouse.pl
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://novahouse-chatbot-api.appspot.com
    description: Production server (GCP App Engine)
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Admin API key for protected endpoints
  schemas:
...
    Lead:
      type: object
      required:
        - name
        - email
        - phone
        - service_type
      properties:
        id:
          type: integer
          description: Lead ID
          example: 42
        name:
          type: string
          description: Lead full name
          example: "Jan Kowalski"
        email:
          type: string
          format: email
          description: Lead email address
          example: "jan.kowalski@example.com"
        phone:
          type: string
          description: Lead phone number
          example: "+48 123 456 789"
        service_type:
          type: string
          description: Requested service type
          enum: [renovation, finishing, design, consultation, other]
          example: "renovation"
        package:
          type: string
          description: Service package (S/M/L)
          enum: [S, M, L]
          example: "M"
        budget:
          type: number
          format: float
          description: Estimated budget in PLN
          example: 50000.0
        timeline:
          type: string
          description: Expected timeline
          example: "3-6 miesięcy"
        status:
          type: string
          description: Lead status
          enum: [new, contacted, qualified, converted, lost]
          example: "new"
        qualification_score:
          type: integer
          description: Qualification score (0-100)
          example: 85
        notes:
          type: string
          description: Additional notes
          example: "Zainteresowany kompleksowym remontem kuchni i łazienki"
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2025-01-15T10:30:00Z"
        rodo_consent:
          type: boolean
          description: RODO consent given
          example: true
...
paths:
...
  /api/leads:
    get:
      tags: [leads]
      summary: Get all leads
      description: Returns paginated list of leads with optional filtering
      security:
        - ApiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Leads list
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer

    post:
      tags: [leads]
      summary: Create new lead
      description: Creates new lead and sends email notifications to admin and customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lead'
      responses:
        '201':
          description: Lead created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Lead created successfully"
                  lead_id:
                    type: integer
                    example: 42
                  monday_item_id:
                    type: string
                    example: "12345678"
  /api/leads/filter:
    post:
      tags: [leads]
      summary: Filter leads
      description: Advanced lead filtering by status, date range, package, qualification score
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [new, contacted, qualified, converted, lost]
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                package:
                  type: string
                  enum: [S, M, L]
                min_score:
                  type: integer
                  minimum: 0
                  maximum: 100
                page:
                  type: integer
                  default: 1
                per_page:
                  type: integer
                  default: 20
      responses:
        '200':
          description: Filtered leads
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items:
                      $ref: '#/components/schemas/Lead'
                  total:
                    type: integer
  /api/leads/export:
    post:
      tags: [leads]
      summary: Export leads to CSV
      description: Export filtered leads to CSV file
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        '200':
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
                format: binary
  /api/leads/bulk-update:
    post:
      tags: [leads]
      summary: Bulk update leads
      description: Update status for multiple leads at once
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - lead_ids
                - status
              properties:
                lead_ids:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3, 4, 5]
                status:
                  type: string
                  enum: [new, contacted, qualified, converted, lost]
                  example: "contacted"
      responses:
        '200':
          description: Bulk update successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Updated 5 leads successfully"
                  updated_count:
                    type: integer
                    example: 5
...